requirements.txt
python-dotenv
streamlit
streamlit_navigation_bar
langchain
langchain_core
langchain_cohere